% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{risk_demo_data}
\alias{risk_demo_data}
\title{Sample risk dataset (risk_demo_data)}
\format{
A dataframe with 8 rows and seven variables:
\describe{
\item{ID}{ID is the hazard ID}
\item{Title}{Title is the hazard title}
\item{RiskLevel}{RiskLevel is the risk level}
\item{Type}{Type is the type of hazard}
\item{SeverityName}{SeverityName is the severity category}
\item{Severity}{Severity is the number associated with the severity category. ONLY whole numbers 1-5 allowed}
\item{Probability}{Probability is the number associated with the probability category. ONLY whole numbers 1-5 allowed}
}
}
\usage{
risk_demo_data
}
\description{
risk_demo_data is a sample risk dataset. The riskmatrix() input must be in
the exact same format as risk_demo_data.
}
\details{
This data frame contains a demo dataset for the safety package.
}
\keyword{datasets}
