% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmatrix.R
\name{riskmatrix}
\alias{riskmatrix}
\title{Safety riskmatrix function}
\usage{
riskmatrix(risk)
}
\arguments{
\item{risk}{A dataframe describing the failure conditions. Column names should
be: c("ID", "Title", "Risk", "Type","Severity.Name","Severity", "Probability")}
}
\value{
An interactive risk matrix called "risk_plot"
\item{risk_plot}{A risk matrix plot visualizing the risk of the hazard dataframe}
}
\description{
This function can be used to plot an interactive risk matrix for risk assessment.
riskmatrix() plots a 5x5 matrix with Severity of Outcome as the x-axis
and Probability of Occurrence as the y-axis. The matrix is separated into a
grid with each block representing a certain risk category. Blocks are
color-coded according to risk level. Red represents the highest risk and
green represents the lowest risk. When hovered over, each block and hazard
will display a text box explaining the risk parameters.
}
\examples{
risk <- risk_demo_data
riskmatrix(risk)

}
